added error messages to login and sign up pages - also, refactored personal best and account page logic
