+++ frontend/src/pages/login/login.css
@@ -104,3 +104,11 @@
     font-family: var(--font-family);
     margin-left: 10px;
 }
+
+.ft__login__error {
+    color: red;
+    font-family: var(--font-family);
+    margin-top: 20px;
+    text-align: center;
+    font-size: 20px;
+}
